Rework nav and routes + add scaffolding for network forms
