+++ ui/src/app/settings/containers/Network/DnsForm/DnsForm.js
@@ -1,24 +1,23 @@
 import React from "react";
 import { useSelector } from "react-redux";
 
-import selectors from "app/settings/selectors";
+import config from "app/settings/selectors/config";
 import Col from "app/base/components/Col";
 import Loader from "app/base/components/Loader";
 import Row from "app/base/components/Row";
-import StorageForm from "app/settings/containers/StorageForm";
 
-const Storage = () => {
-  const loaded = useSelector(selectors.config.loaded);
-  const loading = useSelector(selectors.config.loading);
+const DnsForm = () => {
+  const loaded = useSelector(config.loaded);
+  const loading = useSelector(config.loading);
 
   return (
     <Row>
       <Col size={6}>
         {loading && <Loader text="Loading..." />}
-        {loaded && <StorageForm />}
+        {loaded && <h4>DNS</h4>}
       </Col>
     </Row>
   );
 };
 
-export default Storage;
+export default DnsForm;
