+++ ui/src/app/settings/containers/Network/DnsForm/index.js
@@ -1,0 +1,1 @@
+export { default } from "./DnsForm";
