+++ ui/src/app/settings/containers/Network/NetworkDiscoveryForm/index.js
@@ -1,0 +1,1 @@
+export { default } from "./NetworkDiscoveryForm";
