+++ ui/src/app/settings/containers/Network/NtpForm/index.js
@@ -1,0 +1,1 @@
+export { default } from "./NtpForm";
