+++ ui/src/app/settings/containers/Network/SyslogForm/index.js
@@ -1,0 +1,1 @@
+export { default } from "./SyslogForm";
