--- ui/src/app/settings/containers/Network/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Network";
