--- ui/src/app/settings/containers/Storage/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Storage";
